fix(play): keep the Outputs/Publish/Plan/Activity lists clear of the toolbar band
Same clipping as #130 in the other mailboxes: their full-height Lists
start at the window top and paint the first section header under the
floating glass toolbar, colliding with the nav title.

Share the measured band (ToolbarBandReader, moved to its own file with an
EnvironmentKey) and inset each List with `.safeAreaPadding(.top, band)` —
`.contentMargins` is ignored by List, the explicit safe-area padding is
what actually moves the first row below the band.
